clio-rag/requirements.txt
docling>=2.0
qdrant-client>=1.9
openai>=1.0
anthropic>=0.25
python-dotenv
